+++ examples/experimental/nope/nope_tutorial/high_level_dsl/defs/group_a.yaml
